safari-534.53-branch/LayoutTests/storage/indexeddb/mozilla/object-store-inline-autoincrement-key-added-on-put-expected.txt

Test IndexedDB adding an autoincremented key and retrieving it successfully

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;
PASS indexedDB == null is false
IDBDatabaseException = window.IDBDatabaseException || window.webkitIDBDatabaseException;
PASS IDBDatabaseException == null is false
indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
test = {
        name: 'inline key; key generator',
        autoIncrement: true,
        storedObject: { name: 'Lincoln' },
        keyName: 'id',
    };
objectStore = db.createObjectStore(test.name, { keyPath: test.keyName, autoIncrement: test.autoIncrement });
request = objectStore.add(test.storedObject);
id = event.target.result;
request = objectStore.get(id);
PASS event.target.result.name is test.storedObject.name
PASS event.target.result.id is id
PASS successfullyParsed is true

TEST COMPLETE
